+++ nextgisweb/auth/__init__.py
@@ -2,4 +2,5 @@ from .component import AuthComponent
 from .exception import UserDisabledException
 from .model import Group, OnFindReferencesData, Principal, User
 from .oauth import OAuthAToken, OAuthHelper, OAuthPToken, OnAccessTokenToUser
+from .permission import Permission
 from .policy import AuthProvider, AuthState, OnUserLogin, SecurityPolicy