+++ nextgisweb/auth/migration/435611dc-add-permissions-column.rw.sql
@@ -1,0 +1,3 @@
+/*** { "revision": "435611dc" } ***/
+
+ALTER TABLE auth_principal DROP COLUMN permissions;