+++ nextgisweb/auth/nodepkg/component/index.ts
@@ -1,1 +1,2 @@
 export { PrincipalSelect } from "./principal-select";
+export { PermissionSelect } from "./permission-select";