+++ nextgisweb/auth/nodepkg/component/permission-select/index.ts
@@ -1,2 +1,1 @@
-export { PrincipalSelect } from "./PrincipalSelect";
 export { PermissionSelect } from "./PermissionSelect";